Refactor the mesh conversion routines for MPI
